Add meta-data caching + disable model param to always use the best available

+++ src/main/java/fr/agaspardcilia/chuchot/store/metadata/AudioMetaData.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record AudioMetaData(
         String name,
-        int duration
+        Long duration
 ) implements ItemMetaData {
     @JsonProperty
     @Override
